Refactorm moved three legacy service routines to the new library 'legacyDriver'
